19-03-2023

+++ src/main/java/com/example/OnlineEventBooking/Service/ClientService.java
@@ -1,5 +1,6 @@
 package com.example.OnlineEventBooking.Service;
 
+import com.example.OnlineEventBooking.Entity.Client;
 import com.example.OnlineEventBooking.Model.ClientModel;
 import com.example.OnlineEventBooking.Repository.ClientRepository;
 import jakarta.transaction.Transactional;
@@ -11,16 +12,51 @@ public class ClientService {
 
     @Autowired
     private ClientRepository clientRepository;
-
     @Transactional
-    public ClientModel saveClient(ClientModel clientModel){
-        return  clientModel.assemble(clientRepository.save(clientModel.dissamble()));
+    public String saveClient(ClientModel clientModel){
+        String result;
+        if(searchClient(clientModel.dissamble().getId())){
+            result="Username already exists. Please find another one";
+        }else {
+            upsert(clientModel);
+            result="Save Successfully";
+        }
+        return  result;
     }
-
+    @Transactional
+    public String updateClient(ClientModel clientModel){
+        String result;
+        if(searchClient(clientModel.dissamble().getId())){
+            upsert(clientModel);
+            result="Updated";
+        }else{
+            result="Oops! User not found.";
+        }
+        return result;
+    }
+    @Transactional
+    public boolean searchClient(String  clientId){
+        return clientRepository.existsById(clientId);
+    }
+    @Transactional
+    public ClientModel upsert(ClientModel clientModel){
+        return clientModel.assemble(clientRepository.save(clientModel.dissamble()));
+    }
+    @Transactional
+    public ClientModel getClient(String id){
+        ClientModel clientModel=new ClientModel();
+        return clientModel.assemble(clientRepository.findClientById(id));
+    }
     @Transactional
     public String deleteClient(String clientId){
-        clientRepository.deleteById(clientId);
-        return "Deleted";
+        String result;
+        if(searchClient(clientId)) {
+            clientRepository.deleteById(clientId);
+            result="Deleted";
+        }else{
+            result="Don't exists";
+        }
+        return result;
     }
 
 }
